Сборка на MSVS 2022 версии 17.6 Release 64 без предупреждений.

+++ src/SimpleKafka1C.cpp
@@ -964,7 +964,7 @@ int32_t SimpleKafka1C::getMessageBrokerID()
 
 int32_t SimpleKafka1C::getMessageTimestamp()
 {
-	return this->timestamp / 1000;
+	return (int32_t)this->timestamp / 1000;
 }
 
 int32_t SimpleKafka1C::getMessagePartition()
